Full Solar System First Draft
This is the first code push to the repo. We have used various classes to store in the data from solarsystem.dat file. This is a class project and our end goal is to visually represent a full solar system model. Currently we are stuck with calculating the proper locations of each body in the solar system.